refactor(cli): separate command boundaries
