+++ config/architecture-budgets.json
@@ -6,7 +6,6 @@
     "apps/portal/src/main.tsx": 2251,
     "crates/connect-agent/src/mirrors.rs": 1594,
     "crates/connect-agent/src/server.rs": 1967,
-    "crates/connect-cli/src/lib.rs": 1815,
     "crates/connect-mirror/src/lib.rs": 1915,
     "packages/sync/src/adoption.ts": 1127,
     "packages/sync/src/index.ts": 1116,
