+++ crates/connect-cli/src/lib.rs
@@ -1,4 +1,15 @@
+mod control;
+mod daemon;
+mod login;
+mod output;
 mod service;
+
+use control::{control_command, successful_result};
+use daemon::{
+    connect_paths, doctor, execute_daemon_command, restart_daemon, send, ConnectPaths, DaemonTarget,
+};
+use login::login;
+use output::{print_result, render_connect_profile, render_data_result, OutputKind};
 
 use clap::{Parser, Subcommand};
 use mdbase_command::Command as DataCommand;
@@ -591,122 +602,6 @@ async fn execute(args: Args) -> Result<i32, CliError> {
     }
 }
 
-fn render_connect_profile(
-    report: &mdbase_connect_core::profiling::ProfileReport,
-    output: Option<&Path>,
-    output_json: bool,
-) -> Result<(), CliError> {
-    let serialized = serde_json::to_string_pretty(report)
-        .map_err(|error| CliError::internal(error.to_string()))?;
-    if let Some(output) = output {
-        if let Some(parent) = output.parent() {
-            std::fs::create_dir_all(parent)
-                .map_err(|error| CliError::internal(error.to_string()))?;
-        }
-        std::fs::write(output, format!("{serialized}\n"))
-            .map_err(|error| CliError::internal(error.to_string()))?;
-    }
-    if output_json {
-        println!("{serialized}");
-    } else {
-        println!("Connect profile (read-only)");
-        println!(
-            "{:<26} {:>6} {:>11} {:>11} {:>11}",
-            "operation", "runs", "mean", "p95", "max"
-        );
-        for operation in &report.operations {
-            println!(
-                "{:<26} {:>6} {:>8.2} ms {:>8.2} ms {:>8.2} ms",
-                operation.name,
-                operation.iterations,
-                operation.mean_ms,
-                operation.p95_ms,
-                operation.max_ms,
-            );
-        }
-    }
-    Ok(())
-}
-
-fn render_data_result(value: &Value, pretty: bool, diagnostic: bool) -> Result<(), CliError> {
-    let pretty = pretty || std::io::stdout().is_terminal();
-    let rendered = if pretty {
-        serde_json::to_string_pretty(value)
-    } else {
-        serde_json::to_string(value)
-    }
-    .map_err(|error| CliError::internal(error.to_string()))?;
-    if diagnostic {
-        eprintln!("{rendered}");
-    } else {
-        println!("{rendered}");
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DaemonTarget {
-    InstalledService,
-    IsolatedProfile,
-}
-
-impl DaemonTarget {
-    const fn from_overrides(state_dir: bool, endpoint: bool) -> Self {
-        if state_dir || endpoint {
-            Self::IsolatedProfile
-        } else {
-            Self::InstalledService
-        }
-    }
-
-    const fn label(self) -> &'static str {
-        match self {
-            Self::InstalledService => "installed_service",
-            Self::IsolatedProfile => "isolated_profile",
-        }
-    }
-
-    fn require_installed_service(self, operation: &str) -> Result<(), CliError> {
-        if self == Self::InstalledService {
-            Ok(())
-        } else {
-            Err(CliError::usage(format!(
-                "`mdbase connect daemon {operation}` targets only the default installed service; \
-                 remove --state-dir and --endpoint, or use `daemon run` for an isolated profile."
-            )))
-        }
-    }
-}
-
-struct ConnectPaths {
-    state_dir: PathBuf,
-    endpoint: String,
-    target: DaemonTarget,
-}
-
-fn connect_paths(
-    state_dir: Option<PathBuf>,
-    endpoint: Option<String>,
-) -> Result<ConnectPaths, CliError> {
-    let target = DaemonTarget::from_overrides(state_dir.is_some(), endpoint.is_some());
-    let state_dir = state_dir
-        .map(Ok)
-        .unwrap_or_else(default_state_dir)
-        .map_err(|error| CliError::internal(error.to_string()))?;
-    let state_dir = std::path::absolute(state_dir).map_err(|error| {
-        CliError::internal(format!("Could not resolve the state path: {error}"))
-    })?;
-    let endpoint = endpoint
-        .map(resolve_control_endpoint)
-        .transpose()?
-        .unwrap_or_else(|| default_control_endpoint(&state_dir));
-    Ok(ConnectPaths {
-        state_dir,
-        endpoint,
-        target,
-    })
-}
-
 async fn execute_connect(
     command: ConnectCommand,
     state_dir: Option<PathBuf>,
@@ -833,1077 +728,5 @@ async fn execute_connect(
     }
 }
 
-#[cfg(unix)]
-fn resolve_control_endpoint(endpoint: String) -> Result<String, CliError> {
-    std::path::absolute(&endpoint)
-        .map(|path| path.to_string_lossy().into_owned())
-        .map_err(|error| {
-            CliError::internal(format!("Could not resolve the control endpoint: {error}"))
-        })
-}
-
-#[cfg(windows)]
-fn resolve_control_endpoint(endpoint: String) -> Result<String, CliError> {
-    Ok(endpoint)
-}
-
-fn control_command(command: ConnectCommand) -> Result<(ControlCommand, OutputKind), CliError> {
-    let pair = match command {
-        ConnectCommand::Status => (ControlCommand::Status, OutputKind::Status),
-        ConnectCommand::Whoami => (ControlCommand::AccessSnapshot, OutputKind::Account),
-        ConnectCommand::Ping => (ControlCommand::Ping, OutputKind::Generic),
-        ConnectCommand::Collection(CollectionCommand::List) => {
-            (ControlCommand::CollectionList, OutputKind::Collections)
-        }
-        ConnectCommand::Collection(CollectionCommand::Add { path }) => (
-            ControlCommand::CollectionAdd(CollectionPathParams {
-                path: path_string(path)?,
-            }),
-            OutputKind::Collection,
-        ),
-        ConnectCommand::Collection(CollectionCommand::AddCopy { path }) => (
-            ControlCommand::CollectionAddCopy(CollectionPathParams {
-                path: path_string(path)?,
-            }),
-            OutputKind::Collection,
-        ),
-        ConnectCommand::Collection(CollectionCommand::Create { path, name }) => (
-            ControlCommand::CollectionCreate(CollectionCreateParams {
-                path: path_string(path)?,
-                name,
-            }),
-            OutputKind::Collection,
-        ),
-        ConnectCommand::Collection(CollectionCommand::Remove { collection_id }) => (
-            ControlCommand::CollectionRemove(CollectionIdParams { collection_id }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Collection(CollectionCommand::Validate { collection_id }) => (
-            ControlCommand::CollectionValidate(CollectionIdParams { collection_id }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Collection(CollectionCommand::TransferAuthority {
-            collection_id,
-            target,
-        }) => (
-            ControlCommand::CollectionTransferAuthority(CollectionAuthorityTransferParams {
-                collection_id,
-                target: match target {
-                    CliAuthorityTarget::Remote => AuthorityTarget::Remote,
-                },
-            }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Mirror(MirrorCommand::List) => {
-            (ControlCommand::MirrorList, OutputKind::Mirrors)
-        }
-        ConnectCommand::Mirror(MirrorCommand::Add {
-            collection_id,
-            path,
-            name,
-            read_only,
-            two_way: _,
-        }) => (
-            ControlCommand::MirrorAdd(MirrorAddParams {
-                collection_id,
-                path: path_string(path)?,
-                mode: if read_only {
-                    SyncReplicaMode::ReadOnly
-                } else {
-                    SyncReplicaMode::ReadWrite
-                },
-                name,
-            }),
-            OutputKind::Mirror,
-        ),
-        ConnectCommand::Mirror(MirrorCommand::Sync { replica_id }) => (
-            ControlCommand::MirrorSync(MirrorIdParams { replica_id }),
-            OutputKind::Mirror,
-        ),
-        ConnectCommand::Mirror(MirrorCommand::Resolve {
-            replica_id,
-            record_id,
-            r#use,
-        }) => (
-            ControlCommand::MirrorResolve(MirrorResolveParams {
-                replica_id,
-                record_id,
-                resolution: match r#use {
-                    CliMirrorResolution::Local => MirrorResolution::Local,
-                    CliMirrorResolution::Hosted => MirrorResolution::Remote,
-                },
-            }),
-            OutputKind::Mirror,
-        ),
-        ConnectCommand::Mirror(MirrorCommand::Remove { replica_id, yes }) => {
-            if !yes {
-                return Err(CliError::usage(
-                    "Mirror removal revokes its remote replica. Re-run with --yes to confirm.",
-                ));
-            }
-            (
-                ControlCommand::MirrorRemove(MirrorIdParams { replica_id }),
-                OutputKind::Generic,
-            )
-        }
-        ConnectCommand::Mirror(MirrorCommand::Promote { .. }) => {
-            unreachable!("authority promotion is an interactive CLI flow")
-        }
-        ConnectCommand::Hosted(HostedCommand::List) => {
-            (ControlCommand::HostedSnapshot, OutputKind::Generic)
-        }
-        ConnectCommand::Hosted(HostedCommand::Create { name }) => (
-            ControlCommand::HostedCollectionCreate(HostedCollectionCreateParams { name }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Hosted(HostedCommand::Rename {
-            collection_id,
-            name,
-        }) => (
-            ControlCommand::HostedCollectionRename(HostedCollectionRenameParams {
-                collection_id,
-                name,
-            }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Hosted(HostedCommand::Delete { collection_id, yes }) => {
-            if !yes {
-                return Err(CliError::usage(
-                    "Hosted collection deletion is permanent. Re-run with --yes to confirm.",
-                ));
-            }
-            (
-                ControlCommand::HostedCollectionDelete(CollectionIdParams { collection_id }),
-                OutputKind::Generic,
-            )
-        }
-        ConnectCommand::Operation {
-            collection_id,
-            operation,
-            input,
-        } => (
-            ControlCommand::CollectionOperation(CollectionOperationParams {
-                collection_id,
-                operation,
-                input: serde_json::from_str::<Value>(&input).map_err(|error| {
-                    CliError::usage(format!("--input is not valid JSON: {error}"))
-                })?,
-            }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Access(AccessCommand::List) => {
-            (ControlCommand::AccessSnapshot, OutputKind::Access)
-        }
-        ConnectCommand::Access(AccessCommand::Pause) => (
-            ControlCommand::AccessPause(AccessPauseParams { paused: true }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Access(AccessCommand::Resume) => (
-            ControlCommand::AccessPause(AccessPauseParams { paused: false }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Access(AccessCommand::Approve {
-            request_id,
-            collection_id,
-            operations,
-        }) => (
-            ControlCommand::AuthorizationApprove(AuthorizationApproveParams {
-                request_id,
-                collection_id,
-                operations,
-            }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Access(AccessCommand::Deny { request_id }) => (
-            ControlCommand::AuthorizationDeny(AuthorizationIdParams { request_id }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Access(AccessCommand::Update {
-            grant_id,
-            operations,
-        }) => (
-            ControlCommand::GrantUpdate(GrantUpdateParams {
-                grant_id,
-                operations,
-            }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Access(AccessCommand::Revoke { grant_id }) => (
-            ControlCommand::GrantRevoke(GrantIdParams { grant_id }),
-            OutputKind::Generic,
-        ),
-        ConnectCommand::Activity { limit } => (
-            ControlCommand::ActivityList(ActivityListParams {
-                limit: limit.clamp(1, 500),
-            }),
-            OutputKind::Activity,
-        ),
-        ConnectCommand::Daemon(_)
-        | ConnectCommand::Doctor
-        | ConnectCommand::Login { .. }
-        | ConnectCommand::Logout
-        | ConnectCommand::Paths => {
-            unreachable!("handled before control dispatch")
-        }
-    };
-    Ok(pair)
-}
-
-fn path_string(path: PathBuf) -> Result<String, CliError> {
-    path.into_os_string()
-        .into_string()
-        .map_err(|_| CliError::usage("Collection paths must be valid UTF-8."))
-}
-
-fn successful_result(response: ControlResponse) -> Result<Value, CliError> {
-    if response.protocol_version != LOCAL_CONTROL_PROTOCOL_VERSION {
-        return Err(CliError::internal(format!(
-            "The daemon uses unsupported local protocol {}; expected {}.",
-            response.protocol_version, LOCAL_CONTROL_PROTOCOL_VERSION
-        )));
-    }
-    if response.ok {
-        return Ok(response.result.unwrap_or(Value::Null));
-    }
-    let error = response
-        .error
-        .unwrap_or(mdbase_connect_protocol::ControlError {
-            code: "request_failed".to_string(),
-            message: "The Connect daemon rejected the request.".to_string(),
-            details: None,
-        });
-    Err(CliError {
-        code: error.code,
-        message: error.message,
-        exit_code: 1,
-    })
-}
-
-async fn execute_daemon_command(
-    command: DaemonCommand,
-    state_dir: &Path,
-    endpoint: &str,
-    target: DaemonTarget,
-) -> Result<Value, CliError> {
-    let executable = std::env::current_exe().map_err(|error| {
-        CliError::internal(format!("Could not locate this executable: {error}"))
-    })?;
-    match command {
-        DaemonCommand::Install => {
-            target.require_installed_service("install")?;
-            let installed = service::installed();
-            let running = send(endpoint, ControlRequest::new(ControlCommand::Ping))
-                .await
-                .is_ok_and(|response| response.ok);
-            if installed {
-                match service::stop() {
-                    Ok(()) if running => wait_until_stopped(endpoint).await?,
-                    Ok(()) => {}
-                    Err(error) if running => return Err(CliError::internal(error)),
-                    Err(_) => {}
-                }
-            } else if running {
-                let response = send(
-                    endpoint,
-                    ControlRequest::new(ControlCommand::DaemonShutdown),
-                )
-                .await?;
-                if !response.ok {
-                    return Err(CliError::internal(
-                        response
-                            .error
-                            .map(|error| error.message)
-                            .unwrap_or_else(|| "The daemon refused to stop.".to_string()),
-                    ));
-                }
-                wait_until_stopped(endpoint).await?;
-            }
-            service::install(&executable, state_dir)
-                .map_err(CliError::internal)
-                .map(|_| serde_json::json!({"installed": true}))
-        }
-        DaemonCommand::Uninstall => {
-            target.require_installed_service("uninstall")?;
-            service::uninstall()
-                .map_err(CliError::internal)
-                .map(|_| serde_json::json!({"installed": false}))
-        }
-        DaemonCommand::Start => {
-            if send(endpoint, ControlRequest::new(ControlCommand::Ping))
-                .await
-                .is_ok_and(|response| response.ok)
-            {
-                return Ok(serde_json::json!({
-                    "started": false,
-                    "already_running": true
-                }));
-            }
-            if target == DaemonTarget::InstalledService && service::installed() {
-                service::start().map_err(CliError::internal)?;
-            } else {
-                service::spawn_detached(&executable, state_dir, endpoint)
-                    .map_err(CliError::internal)?;
-            }
-            wait_until_ready(endpoint).await?;
-            Ok(serde_json::json!({"started": true}))
-        }
-        DaemonCommand::Stop => {
-            if target == DaemonTarget::InstalledService && service::installed() {
-                service::stop().map_err(CliError::internal)?;
-            } else {
-                let response = send(
-                    endpoint,
-                    ControlRequest::new(ControlCommand::DaemonShutdown),
-                )
-                .await?;
-                if !response.ok {
-                    return Err(CliError::internal(
-                        response
-                            .error
-                            .map(|error| error.message)
-                            .unwrap_or_else(|| "The daemon refused to stop.".to_string()),
-                    ));
-                }
-            }
-            wait_until_stopped(endpoint).await?;
-            Ok(serde_json::json!({"stopped": true}))
-        }
-        DaemonCommand::Restart => {
-            restart_daemon(state_dir, endpoint, target).await?;
-            Ok(serde_json::json!({"restarted": true}))
-        }
-        DaemonCommand::Status => {
-            let installed = target == DaemonTarget::InstalledService && service::installed();
-            let response = send(endpoint, ControlRequest::new(ControlCommand::Status)).await;
-            Ok(match response {
-                Ok(response) if response.ok => serde_json::json!({
-                    "target": target.label(),
-                    "installed": installed,
-                    "running": true,
-                    "status": response.result
-                }),
-                _ => serde_json::json!({
-                    "target": target.label(),
-                    "installed": installed,
-                    "running": false
-                }),
-            })
-        }
-        DaemonCommand::Logs { lines, follow } => {
-            target.require_installed_service("logs")?;
-            service::logs(state_dir, lines, follow).map_err(CliError::internal)?;
-            Ok(serde_json::json!({"shown": true}))
-        }
-        DaemonCommand::Run { .. } => unreachable!("foreground daemon handled before dispatch"),
-    }
-}
-
-async fn doctor(state_dir: &Path, endpoint: &str, target: DaemonTarget) -> Value {
-    let existed = state_dir.exists();
-    let state_directory = match create_private_state_dir(state_dir) {
-        Ok(()) if existed => "available",
-        Ok(()) => "created",
-        Err(_) => "unavailable",
-    };
-    let response = send(endpoint, ControlRequest::new(ControlCommand::Status)).await;
-    let (daemon, status) = match response {
-        Ok(response) if response.ok => ("ready", response.result),
-        _ => ("unavailable", None),
-    };
-    serde_json::json!({
-        "healthy": state_directory != "unavailable" && daemon == "ready",
-        "state_directory": {
-            "path": state_dir,
-            "state": state_directory
-        },
-        "daemon": {
-            "target": target.label(),
-            "endpoint": endpoint,
-            "state": daemon,
-            "status": status
-        },
-        "service_installed":
-            target == DaemonTarget::InstalledService && service::installed()
-    })
-}
-
-fn create_private_state_dir(state_dir: &Path) -> std::io::Result<()> {
-    std::fs::create_dir_all(state_dir)?;
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        std::fs::set_permissions(state_dir, std::fs::Permissions::from_mode(0o700))?;
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone, Copy)]
-enum OutputKind {
-    Status,
-    Daemon,
-    Doctor,
-    Collections,
-    Collection,
-    Access,
-    Activity,
-    Account,
-    Mirrors,
-    Mirror,
-    Generic,
-}
-
-fn print_result(json: bool, kind: OutputKind, value: &Value) -> Result<(), CliError> {
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(value)
-                .map_err(|error| CliError::internal(error.to_string()))?
-        );
-        return Ok(());
-    }
-    println!("{}", render_human(kind, value));
-    Ok(())
-}
-
-fn render_human(kind: OutputKind, value: &Value) -> String {
-    match kind {
-        OutputKind::Status => {
-            let state = value["state"]
-                .as_str()
-                .unwrap_or("unknown")
-                .replace('_', " ");
-            let collections = value["registered_collections"].as_u64().unwrap_or(0);
-            let paused = value["paused"].as_bool().unwrap_or(false);
-            format!(
-                "{}\nCollections: {}\nAccess: {}",
-                sentence_case(&state),
-                collections,
-                if paused { "paused" } else { "available" }
-            )
-        }
-        OutputKind::Daemon => {
-            if value["running"] == Value::Bool(true) {
-                "Daemon is running.".to_string()
-            } else if value["installed"] == Value::Bool(true) {
-                "Daemon is installed but not running.".to_string()
-            } else if value["installed"] == Value::Bool(false) {
-                "Daemon is not installed.".to_string()
-            } else if value["shown"] == Value::Bool(true) {
-                String::new()
-            } else {
-                value
-                    .as_object()
-                    .and_then(|object| object.keys().next())
-                    .map(|key| format!("Daemon {}.", key.replace('_', " ")))
-                    .unwrap_or_else(|| "Done.".to_string())
-            }
-        }
-        OutputKind::Doctor => {
-            let healthy = value["healthy"].as_bool().unwrap_or(false);
-            let state = value["state_directory"]["state"]
-                .as_str()
-                .unwrap_or("unknown");
-            let daemon = value["daemon"]["state"].as_str().unwrap_or("unknown");
-            format!(
-                "{}\nState directory: {}\nDaemon: {}",
-                if healthy {
-                    "Connect is healthy."
-                } else {
-                    "Connect needs attention."
-                },
-                state,
-                daemon
-            )
-        }
-        OutputKind::Collections => render_rows(
-            value.as_array().map(Vec::as_slice).unwrap_or(&[]),
-            &["NAME", "STATE", "ID", "PATH"],
-            |item| {
-                vec![
-                    text(item, "display_name"),
-                    if item["enabled"].as_bool().unwrap_or(false) {
-                        "available".to_string()
-                    } else {
-                        "paused".to_string()
-                    },
-                    text(item, "id"),
-                    text(item, "path"),
-                ]
-            },
-            "No computer-owned collections.",
-        ),
-        OutputKind::Collection => {
-            format!(
-                "{}\n{}\n{}",
-                text(value, "display_name"),
-                text(value, "id"),
-                text(value, "path")
-            )
-        }
-        OutputKind::Access => {
-            let requests = value["pending_authorizations"]
-                .as_array()
-                .map(Vec::as_slice)
-                .unwrap_or(&[]);
-            let grants = value["grants"].as_array().map(Vec::as_slice).unwrap_or(&[]);
-            format!(
-                "Pending requests: {}\nActive grants: {}",
-                requests.len(),
-                grants.len()
-            )
-        }
-        OutputKind::Activity => render_rows(
-            value.as_array().map(Vec::as_slice).unwrap_or(&[]),
-            &["TIME", "ACTION", "OUTCOME"],
-            |item| {
-                vec![
-                    text(item, "occurred_at"),
-                    text(item, "action"),
-                    text(item, "outcome"),
-                ]
-            },
-            "No recent activity.",
-        ),
-        OutputKind::Account => {
-            if value["configured"] == Value::Bool(false) {
-                return "This computer is not connected to an account.".to_string();
-            }
-            let account = value.get("account").unwrap_or(value);
-            let user = account["user_name"].as_str().unwrap_or("Connected");
-            let email = account["user_email"].as_str().unwrap_or("");
-            let computer = account["connector_name"].as_str().unwrap_or("");
-            [user, email, computer]
-                .into_iter()
-                .filter(|value| !value.is_empty())
-                .collect::<Vec<_>>()
-                .join("\n")
-        }
-        OutputKind::Mirrors => render_rows(
-            value.as_array().map(Vec::as_slice).unwrap_or(&[]),
-            &["NAME", "STATE", "MODE", "ID", "PATH"],
-            |item| {
-                vec![
-                    text(item, "name"),
-                    text(item, "state").replace('_', " "),
-                    text(item, "mode").replace('_', " "),
-                    text(item, "replica_id"),
-                    text(item, "path"),
-                ]
-            },
-            "No hosted collection mirrors.",
-        ),
-        OutputKind::Mirror => {
-            let error = value["error"].as_str();
-            format!(
-                "{}\n{}\n{}\n{}{}",
-                text(value, "name"),
-                text(value, "state").replace('_', " "),
-                text(value, "path"),
-                text(value, "replica_id"),
-                error.map(|error| format!("\n{error}")).unwrap_or_default()
-            )
-        }
-        OutputKind::Generic => {
-            if value.is_null() || value == &serde_json::json!({}) {
-                "Done.".to_string()
-            } else {
-                serde_json::to_string_pretty(value).unwrap_or_else(|_| "Done.".to_string())
-            }
-        }
-    }
-}
-
-fn text(value: &Value, field: &str) -> String {
-    value[field]
-        .as_str()
-        .map(str::to_string)
-        .unwrap_or_else(|| value[field].to_string().trim_matches('"').to_string())
-}
-
-fn sentence_case(value: &str) -> String {
-    let mut characters = value.chars();
-    match characters.next() {
-        Some(first) => first.to_uppercase().collect::<String>() + characters.as_str(),
-        None => String::new(),
-    }
-}
-
-fn render_rows(
-    values: &[Value],
-    headings: &[&str],
-    row: impl Fn(&Value) -> Vec<String>,
-    empty: &str,
-) -> String {
-    if values.is_empty() {
-        return empty.to_string();
-    }
-    let rows = values.iter().map(row).collect::<Vec<_>>();
-    let mut widths = headings
-        .iter()
-        .map(|heading| heading.len())
-        .collect::<Vec<_>>();
-    for values in &rows {
-        for (index, value) in values.iter().enumerate() {
-            widths[index] = widths[index].max(value.chars().count());
-        }
-    }
-    let format_row = |values: &[String]| {
-        values
-            .iter()
-            .enumerate()
-            .map(|(index, value)| format!("{value:<width$}", width = widths[index]))
-            .collect::<Vec<_>>()
-            .join("  ")
-            .trim_end()
-            .to_string()
-    };
-    let heading_values = headings
-        .iter()
-        .map(|heading| heading.to_string())
-        .collect::<Vec<_>>();
-    std::iter::once(format_row(&heading_values))
-        .chain(rows.iter().map(|values| format_row(values)))
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-#[derive(Debug, serde::Deserialize)]
-struct PairingBegin {
-    pairing_id: Uuid,
-    pairing_secret: String,
-    verification_uri: String,
-    expires_in: u64,
-}
-
-#[derive(Debug, serde::Deserialize)]
-struct PairingExchange {
-    status: String,
-    token: Option<String>,
-    connector: Option<Value>,
-    error: Option<PairingError>,
-}
-
-#[derive(Debug, serde::Deserialize)]
-struct PairingError {
-    message: Option<String>,
-}
-
-async fn login(
-    state_dir: &Path,
-    endpoint: &str,
-    target: DaemonTarget,
-    server: &str,
-    requested_name: Option<&str>,
-    no_open: bool,
-) -> Result<Value, CliError> {
-    let configuration =
-        CloudConfiguration::new(server).map_err(|error| CliError::usage(error.to_string()))?;
-    let connector_name = requested_name
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_string)
-        .unwrap_or_else(default_computer_name);
-    if connector_name.chars().count() > 100 {
-        return Err(CliError::usage(
-            "Computer name must be between 1 and 100 characters.",
-        ));
-    }
-    let client = reqwest::Client::new();
-    let response = client
-        .post(format!("{}/v1/pairing-requests", configuration.server_url))
-        .json(&serde_json::json!({"connector_name": connector_name}))
-        .send()
-        .await
-        .map_err(|error| {
-            CliError::unavailable(format!("Could not reach the Connect server: {error}"))
-        })?;
-    let status = response.status();
-    let value = response.json::<Value>().await.map_err(|error| {
-        CliError::internal(format!(
-            "Connect returned an invalid pairing response: {error}"
-        ))
-    })?;
-    if !status.is_success() {
-        return Err(CliError {
-            code: value
-                .pointer("/error/code")
-                .and_then(Value::as_str)
-                .unwrap_or("pairing_failed")
-                .to_string(),
-            message: value
-                .pointer("/error/message")
-                .and_then(Value::as_str)
-                .unwrap_or("Computer pairing could not begin.")
-                .to_string(),
-            exit_code: 1,
-        });
-    }
-    let pairing = serde_json::from_value::<PairingBegin>(value).map_err(|error| {
-        CliError::internal(format!("Connect returned invalid pairing details: {error}"))
-    })?;
-    if pairing.expires_in == 0
-        || pairing.expires_in > 86_400
-        || !valid_secret(&pairing.pairing_secret, "pair_")
-    {
-        return Err(CliError::internal(
-            "Connect returned unsafe pairing details.",
-        ));
-    }
-    let verification = url::Url::parse(&pairing.verification_uri)
-        .map_err(|_| CliError::internal("Connect returned an invalid verification address."))?;
-    let expected = url::Url::parse(&configuration.server_url)
-        .map_err(|_| CliError::internal("Configured Connect server is invalid."))?;
-    if verification.origin() != expected.origin()
-        || !verification.username().is_empty()
-        || verification.password().is_some()
-    {
-        return Err(CliError::internal(
-            "Connect returned a verification address on another origin.",
-        ));
-    }
-    eprintln!("Approve this computer in your browser:\n{verification}");
-    if !no_open {
-        service::open_url(verification.as_str()).map_err(CliError::internal)?;
-    }
-    let deadline = std::time::Instant::now() + std::time::Duration::from_secs(pairing.expires_in);
-    loop {
-        if std::time::Instant::now() >= deadline {
-            return Err(CliError {
-                code: "pairing_expired".to_string(),
-                message: "Computer approval expired before it was completed.".to_string(),
-                exit_code: 1,
-            });
-        }
-        let response = client
-            .post(format!(
-                "{}/v1/pairing-requests/{}/exchange",
-                configuration.server_url, pairing.pairing_id
-            ))
-            .bearer_auth(&pairing.pairing_secret)
-            .send()
-            .await;
-        match response {
-            Ok(response) if response.status().as_u16() == 202 => {}
-            Ok(response) => {
-                let status = response.status();
-                let exchange = response.json::<PairingExchange>().await.map_err(|error| {
-                    CliError::internal(format!(
-                        "Connect returned an invalid pairing result: {error}"
-                    ))
-                })?;
-                if !status.is_success() {
-                    return Err(CliError {
-                        code: "pairing_failed".to_string(),
-                        message: exchange
-                            .error
-                            .and_then(|error| error.message)
-                            .unwrap_or_else(|| {
-                                format!("Computer pairing failed with HTTP {status}.")
-                            }),
-                        exit_code: 1,
-                    });
-                }
-                if exchange.status != "paired" {
-                    return Err(CliError::internal(
-                        "Connect returned an invalid pairing state.",
-                    ));
-                }
-                let token = exchange.token.ok_or_else(|| {
-                    CliError::internal("Connect returned no connector credential.")
-                })?;
-                let configured = send(
-                    endpoint,
-                    ControlRequest::new(ControlCommand::AccountConfigure(
-                        mdbase_connect_protocol::AccountConfigureParams {
-                            server_url: configuration.server_url.clone(),
-                            connector_token: token.clone(),
-                        },
-                    )),
-                )
-                .await
-                .is_ok_and(|response| response.ok);
-                if !configured {
-                    configure_cloud(state_dir, &configuration, &token)
-                        .map_err(|error| CliError::internal(error.to_string()))?;
-                }
-                restart_daemon(state_dir, endpoint, target).await?;
-                return Ok(serde_json::json!({
-                    "configured": true,
-                    "server_url": configuration.server_url,
-                    "account": exchange.connector
-                }));
-            }
-            Err(_) => {}
-        }
-        tokio::time::sleep(std::time::Duration::from_millis(1_500)).await;
-    }
-}
-
-async fn restart_daemon(
-    state_dir: &Path,
-    endpoint: &str,
-    target: DaemonTarget,
-) -> Result<(), CliError> {
-    if target == DaemonTarget::InstalledService && service::installed() {
-        service::restart().map_err(CliError::internal)?;
-    } else {
-        let _ = send(
-            endpoint,
-            ControlRequest::new(ControlCommand::DaemonShutdown),
-        )
-        .await;
-        wait_until_stopped(endpoint).await?;
-        let executable = std::env::current_exe().map_err(|error| {
-            CliError::internal(format!("Could not locate this executable: {error}"))
-        })?;
-        service::spawn_detached(&executable, state_dir, endpoint).map_err(CliError::internal)?;
-    }
-    wait_until_ready(endpoint).await
-}
-
-async fn wait_until_stopped(endpoint: &str) -> Result<(), CliError> {
-    for _ in 0..100 {
-        if send(endpoint, ControlRequest::new(ControlCommand::Ping))
-            .await
-            .is_err()
-        {
-            return Ok(());
-        }
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-    }
-    Err(CliError::unavailable(
-        "The Connect daemon did not stop within five seconds.",
-    ))
-}
-
-async fn wait_until_ready(endpoint: &str) -> Result<(), CliError> {
-    for _ in 0..200 {
-        if send(endpoint, ControlRequest::new(ControlCommand::Ping))
-            .await
-            .is_ok_and(|response| {
-                response.ok
-                    && response
-                        .result
-                        .as_ref()
-                        .and_then(|value| value["ready"].as_bool())
-                        .unwrap_or(false)
-            })
-        {
-            return Ok(());
-        }
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-    }
-    Err(CliError::unavailable(
-        "The Connect daemon did not become ready.",
-    ))
-}
-
-fn default_computer_name() -> String {
-    std::env::var("HOSTNAME")
-        .or_else(|_| std::env::var("COMPUTERNAME"))
-        .ok()
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-        .unwrap_or_else(|| "This computer".to_string())
-        .chars()
-        .take(100)
-        .collect()
-}
-
-fn valid_secret(secret: &str, prefix: &str) -> bool {
-    secret.starts_with(prefix) && secret.len() >= 24 && !secret.chars().any(char::is_whitespace)
-}
-
-async fn exchange<S>(stream: S, request: ControlRequest) -> Result<ControlResponse, CliError>
-where
-    S: tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin,
-{
-    let (reader, mut writer) = tokio::io::split(stream);
-    let request_id = request.id;
-    let mut encoded =
-        serde_json::to_vec(&request).map_err(|error| CliError::internal(error.to_string()))?;
-    encoded.push(b'\n');
-    writer
-        .write_all(&encoded)
-        .await
-        .map_err(|error| CliError::unavailable(error.to_string()))?;
-    const MAX_LOCAL_CONTROL_RESPONSE_BYTES: u64 = 32 * 1024 * 1024;
-    let mut reader = BufReader::new(reader.take(MAX_LOCAL_CONTROL_RESPONSE_BYTES + 1));
-    let mut line = Vec::new();
-    let read = reader
-        .read_until(b'\n', &mut line)
-        .await
-        .map_err(|error| CliError::unavailable(error.to_string()))?;
-    if read == 0 {
-        return Err(CliError::unavailable(
-            "The daemon closed without a response.",
-        ));
-    }
-    if line.len() as u64 > MAX_LOCAL_CONTROL_RESPONSE_BYTES || line.last() != Some(&b'\n') {
-        return Err(CliError::internal(
-            "The daemon returned an oversized or incomplete response.",
-        ));
-    }
-    line.pop();
-    let response: ControlResponse = serde_json::from_slice(&line).map_err(|error| {
-        CliError::internal(format!("The daemon returned an invalid response: {error}"))
-    })?;
-    if response.id != request_id {
-        return Err(CliError::internal(
-            "The daemon returned a response for a different request.",
-        ));
-    }
-    if response.protocol_version != LOCAL_CONTROL_PROTOCOL_VERSION {
-        return Err(CliError::internal(format!(
-            "The daemon uses unsupported local protocol {}; expected {}.",
-            response.protocol_version, LOCAL_CONTROL_PROTOCOL_VERSION
-        )));
-    }
-    Ok(response)
-}
-
-#[cfg(unix)]
-async fn send(endpoint: &str, request: ControlRequest) -> Result<ControlResponse, CliError> {
-    let timeout = control_request_timeout(&request.command);
-    tokio::time::timeout(timeout, async {
-        let stream = tokio::net::UnixStream::connect(endpoint)
-            .await
-            .map_err(|error| {
-                CliError::unavailable(format!(
-                    "Could not connect to the Connect daemon at {endpoint}: {error}"
-                ))
-            })?;
-        exchange(stream, request).await
-    })
-    .await
-    .map_err(|_| CliError::unavailable("The Connect daemon request timed out."))?
-}
-
-#[cfg(windows)]
-async fn send(endpoint: &str, request: ControlRequest) -> Result<ControlResponse, CliError> {
-    use tokio::net::windows::named_pipe::ClientOptions;
-    let timeout = control_request_timeout(&request.command);
-    tokio::time::timeout(timeout, async {
-        let stream = ClientOptions::new().open(endpoint).map_err(|error| {
-            CliError::unavailable(format!(
-                "Could not connect to the Connect daemon at {endpoint}: {error}"
-            ))
-        })?;
-        exchange(stream, request).await
-    })
-    .await
-    .map_err(|_| CliError::unavailable("The Connect daemon request timed out."))?
-}
-
-fn control_request_timeout(command: &ControlCommand) -> std::time::Duration {
-    let seconds = match command {
-        ControlCommand::MirrorAdd(_)
-        | ControlCommand::MirrorSync(_)
-        | ControlCommand::MirrorResolve(_)
-        | ControlCommand::MirrorPromoteComplete(_)
-        | ControlCommand::CollectionTransferAuthority(_) => 15 * 60,
-        ControlCommand::CollectionOperation(_) => 2 * 60,
-        _ => 30,
-    };
-    std::time::Duration::from_secs(seconds)
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parser_keeps_data_and_connect_namespaces_unambiguous() {
-        let direct =
-            Args::try_parse_from(["mdbase", "--root", "/data/notes", "read", "note.md"]).unwrap();
-        assert!(matches!(
-            direct.command,
-            RootCommand::Data(DataCommand::Read { .. })
-        ));
-
-        let control =
-            Args::try_parse_from(["mdbase", "--json", "connect", "daemon", "status"]).unwrap();
-        assert!(matches!(
-            control.command,
-            RootCommand::Connect {
-                command: ConnectCommand::Daemon(DaemonCommand::Status)
-            }
-        ));
-    }
-
-    #[test]
-    fn parser_rejects_ambiguous_data_targets() {
-        let error = Args::try_parse_from([
-            "mdbase",
-            "--root",
-            "/data/notes",
-            "--collection",
-            "01900000-0000-7000-8000-000000000000",
-            "read",
-            "note.md",
-        ])
-        .unwrap_err();
-        assert_eq!(error.kind(), clap::error::ErrorKind::ArgumentConflict);
-    }
-
-    #[test]
-    fn daemon_targeting_separates_installed_and_isolated_processes() {
-        assert_eq!(
-            DaemonTarget::from_overrides(false, false),
-            DaemonTarget::InstalledService
-        );
-        assert_eq!(
-            DaemonTarget::from_overrides(true, false),
-            DaemonTarget::IsolatedProfile
-        );
-        assert_eq!(
-            DaemonTarget::from_overrides(false, true),
-            DaemonTarget::IsolatedProfile
-        );
-        let error = DaemonTarget::IsolatedProfile
-            .require_installed_service("install")
-            .unwrap_err();
-        assert_eq!(error.code, "invalid_input");
-        assert!(error.message.contains("default installed service"));
-    }
-
-    #[test]
-    fn human_status_is_quiet_and_explicit() {
-        let value = serde_json::json!({
-            "state": "connected",
-            "registered_collections": 2,
-            "paused": false
-        });
-        assert_eq!(
-            render_human(OutputKind::Status, &value),
-            "Connected\nCollections: 2\nAccess: available"
-        );
-    }
-
-    #[test]
-    fn collection_table_has_stable_columns() {
-        let value = serde_json::json!([{
-            "display_name": "Notes",
-            "enabled": true,
-            "id": "01900000-0000-7000-8000-000000000000",
-            "path": "/data/notes"
-        }]);
-        let rendered = render_human(OutputKind::Collections, &value);
-        assert!(rendered.starts_with("NAME"));
-        assert!(rendered.contains("Notes"));
-        assert!(rendered.contains("available"));
-        assert!(!rendered.contains('{'));
-    }
-
-    #[test]
-    fn invalid_operation_json_is_a_usage_error() {
-        let result = control_command(ConnectCommand::Operation {
-            collection_id: Uuid::nil(),
-            operation: "query".to_string(),
-            input: "{".to_string(),
-        });
-        let error = result.unwrap_err();
-        assert_eq!(error.code, "invalid_input");
-        assert_eq!(error.exit_code, 2);
-    }
-}
+mod tests;
